+++ src/main/kotlin/com/example/smartshopping/domain/repository/PurchaseRepository.kt
@@ -1,5 +1,6 @@
-package com.example.smartshopping.domain.purchase
+package com.example.smartshopping.domain.repository
 
+import com.example.smartshopping.entity.PurchaseRecord
 import org.springframework.data.jpa.repository.JpaRepository
 
 interface PurchaseRepository : JpaRepository<PurchaseRecord, Long> {
