+++ src/main/kotlin/com/example/smartshopping/domain/repository/UserRepository.kt
@@ -1,5 +1,6 @@
-package com.example.smartshopping.domain.user
+package com.example.smartshopping.domain.repository
 
+import com.example.smartshopping.entity.User
 import org.springframework.data.jpa.repository.JpaRepository
 import org.springframework.stereotype.Repository
